+++ DbLayer/DataModels/UserReportDb.cs
@@ -11,9 +11,9 @@ namespace DbLayer.DataModels
         public virtual UserDb User { get; set; }
         public int UserId { get; set; }
         public DateTime Date { get; set; }
-        public float WorkHours { get; set; }
-        public float BreakHours { get; set; }
-        public float TimeDifference { get; set; }
+        public double WorkHours { get; set; }
+        public double BreakHours { get; set; }
+        public double TimeDifference { get; set; }
         public string TimeDifferenceReason { get; set; }
         public virtual UserDb ApprovedByUser { get; set; }
         public int ApprovedByUserId { get; set; }
